fix: Corrigindo Bugs de converção dos dados recebeidos pela Api com o modelo das entidades da aplicação     - Corrigindo erro na converção dos dados da api com o modelos de movieEntipy. Erro causado pelo tipo de dado recebido da api, que no campo (vote_average) pode retornar um int ou duoble, e antes era sempre esperado com um int, agora o campo indenpendente se for int ou duoble ele sempre sera convertido para um duoble.

+++ lib/features/movies/home/data/models/movies_page_model.dart
@@ -17,12 +17,12 @@ class MoviesPageModel extends MoviesPageEntipy {
 
   factory MoviesPageModel.fromJson(Map<String, dynamic> pageMoviesJson) {
     return MoviesPageModel(
-      page: pageMoviesJson['page'],
+      page: pageMoviesJson['page'].toInt(),
       movies: (pageMoviesJson['results'] as List)
           .map((movie) => MovieModel.fromJson(movie))
           .toList(),
-      totalPages: pageMoviesJson['total_pages'],
-      totalResults: pageMoviesJson['total_results'],
+      totalPages: pageMoviesJson['total_pages'].toInt(),
+      totalResults: pageMoviesJson['total_results'].toInt(),
     );
   }
 }
